+++ lib/flutter_viewmodel.dart
@@ -6,20 +6,32 @@ import 'package:viewmodel_flutter/states.dart';
 
 abstract class ViewModel extends ChangeNotifier {
   final List<SideEffect> _sideEffects = [];
+  bool _disposed = false;
+
+  bool get isDisposed => _disposed;
 
   void init();
 
-  void update(dynamic state) {
-    for (var sideEffect in _sideEffects) {
-      if (state != null) {
-        if (sideEffect.dependencies.any((dep) => dep == state)) {
-          sideEffect.effect();
-        }
-      } else {
+  bool update([dynamic state]) {
+    if (_disposed) {
+      return false;
+    }
+
+    for (final sideEffect in _sideEffects) {
+      if (state == null || sideEffect.dependencies.any((dep) => dep == state)) {
         sideEffect.effect();
+        if (_disposed) {
+          return false;
+        }
       }
     }
+
+    if (_disposed) {
+      return false;
+    }
+
     notifyListeners();
+    return true;
   }
 
   MutableState<T> mutableStateOf<T>(T state) {
@@ -29,7 +41,7 @@ abstract class ViewModel extends ChangeNotifier {
   MutableStateList<T> mutableStateListOf<T>() {
     return MutableStateList(this);
   }
-  
+
   MutableStateMap<K, V> mutableStateMapOf<K, V>() {
     return MutableStateMap(this);
   }
@@ -38,5 +50,13 @@ abstract class ViewModel extends ChangeNotifier {
 
   void sideEffect(VoidCallback effect, List<dynamic> dependencies) {
     _sideEffects.add(SideEffect(effect, dependencies));
+  }
+
+  @mustCallSuper
+  @override
+  void dispose() {
+    _disposed = true;
+    _sideEffects.clear();
+    super.dispose();
   }
 }